requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-chroma
langchain-community
langchain-text-splitters
python-dotenv
pypdf
